iss 711 p 4.1.1 not tested

+++ Dev/DataModels/Atdi.DataModels.Sdrn.CalcServer.Internal/Iterations/StationCalibrationCalc/MainBlock/Input/CalcPoint.cs
@@ -10,9 +10,9 @@ namespace Atdi.DataModels.Sdrn.CalcServer.Internal.Iterations
 {
 	public struct CalcPoint
     {
-        public float Lon;
+        public int Lon;
 
-        public float Lat;
+        public int Lat;
 
         public int Count;
